Swaps League Statistics helpers to the stat helper
